Fix build failure: signedness mismatch in assert Contributed-by: Michal Vala <mvala@redhat.com>

+++ src/share/vm/gc_implementation/shenandoah/shenandoahHeapRegion.cpp
@@ -340,7 +340,7 @@ size_t ShenandoahHeapRegion::get_gclab_allocs() const {
 void ShenandoahHeapRegion::set_live_data(size_t s) {
   assert(Thread::current()->is_VM_thread(), "by VM thread");
   size_t v = s >> LogHeapWordSize;
-  assert(v < max_jint, "sanity");
+  assert(v < (size_t)max_jint, "sanity");
   _live_data = (jint)v;
 }
